requirements.txt
atlassian==0.0.0
atlassian-python-api==3.28.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
future==0.18.2
idna==3.3
jira==3.4.1
oauthlib==3.2.0
packaging==21.3
py==1.11.0
pyparsing==3.0.9
python-decouple==3.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
retry==0.9.2
singledispatch==3.7.0
six==1.16.0
testrail==0.3.15
testrail-api==1.10.6
typing-extensions==4.3.0
urllib3==1.26.10
wrapt==1.14.1
